refactor : 폴더이동
Styled 컴포넌트들을 style 폴더안으로 이동시켰다.
