+++ src/styles/CardStyled.jsx
