+++ src/styles/SearchStyled.jsx
